requirements.txt
python-dotenv
google-adk==0.3.0
litellm
google-cloud-aiplatform[agent_engines]